+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/simulator/subscription/publisher/PeriodicBusPublisher.java
@@ -13,7 +13,7 @@ public class PeriodicBusPublisher<T> extends SubmissionPublisher<T> {
     private final ScheduledExecutorService scheduler;
     private final ScheduledFuture<?> periodicTask;
 
-    PeriodicBusPublisher(
+    public PeriodicBusPublisher(
             BusSubscriptionManager manager, long busId,
             Supplier<T> supplier, long period, long initialDelay, TimeUnit unit
     ) {
